refactor (step3) domain 패키지를 만들고 domain 클래스를 모음
