+++ src/main/java/kitchenpos/menu/controller/MenuRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.controller;
+package kitchenpos.menu.controller;
 
 import kitchenpos.menu.bo.MenuBo;
-import kitchenpos.model.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
