+++ src/main/java/kitchenpos/menu/domain/MenuDao.java
@@ -5,7 +5,7 @@ import kitchenpos.menu.domain.Menu;
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuRepository {
+public interface MenuDao {
     Menu save(Menu entity);
 
     Optional<Menu> findById(Long id);
