+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.model;
+package kitchenpos.menu.domain;
 
 public class MenuGroup {
     private Long id;
