+++ src/main/java/kitchenpos/menu/domain/ProductDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.dao;
+package kitchenpos.menu.domain;
 
-import kitchenpos.menu.model.Product;
+import kitchenpos.menu.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
