+++ src/main/java/kitchenpos/order/dao/JdbcTemplateOrderDao.java
@@ -1,6 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.order.dao;
 
-import kitchenpos.model.Order;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
