+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.model;
+package kitchenpos.order.domain;
 
 public class OrderLineItem {
     private Long seq;
