+++ src/main/java/kitchenpos/ordertable/bo/TableBo.java
@@ -1,7 +1,9 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.ordertable.bo;
 
 import kitchenpos.order.domain.OrderDao;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableDao;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
