+++ src/main/java/kitchenpos/ordertable/domain/TableGroupDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.ordertable.dao;
+package kitchenpos.ordertable.domain;
 
-import kitchenpos.ordertable.model.TableGroup;
+import kitchenpos.ordertable.domain.TableGroup;
 
 import java.util.List;
 import java.util.Optional;
