+++ src/test/java/kitchenpos/bo/ProductBoTest.java
@@ -1,7 +1,8 @@
 package kitchenpos.bo;
 
-import kitchenpos.dao.ProductDao;
-import kitchenpos.model.Product;
+import kitchenpos.menu.domain.ProductDao;
+import kitchenpos.menu.bo.ProductBo;
+import kitchenpos.menu.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
