+++ src/test/java/kitchenpos/controller/TableGroupRestControllerTest.java
@@ -1,7 +1,8 @@
 package kitchenpos.controller;
 
-import kitchenpos.bo.TableGroupBo;
-import kitchenpos.model.TableGroup;
+import kitchenpos.ordertable.bo.TableGroupBo;
+import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.ordertable.controller.TableGroupRestController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.web.servlet.HttpEncodingAutoConfiguration;
